+++ packages/orq-rc/src/models/operations/streamrunagent.ts
@@ -3949,7 +3949,7 @@ export const AgentToolInputRunTools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   AgentToolInputRunTools
 > = z.object({
-  id: z.string().default("01M32D3D1943D1PDNHG314B9BW"),
+  id: z.string().default("01M337FG89BAT3M1VCT7CE9G1Q"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() =>